(#53) attempt to handle purges on target streams better
If we detect that a Purge probably happened on resume we
will continue from the last message time, this only works
in the case where there is no messages in the stream at all
when we ask for the last sequence.

In other cases we will copy all the data.

Signed-off-by: R.I.Pienaar <rip@devco.net>