(#83) Add Heartbeats to streams
Here we add the ability to configure heartbeat messages for streams.

By adding a new heartbeat section to the configuration file, and specifying
specific streams, stream-replicator will send a message with a unix timestamp as
the body at a predetermined interval. This can be used to monitor availability
when streams are mostly idle.

The heartbeat package will spawn a goroutine for every configured subject and
send a message on interval. The heartbeat publisher is leader aware and adds
relevant prometheus metrics.